broker-test/requirements.txt
kafka-python==2.3.0
pika==1.3.2
python-dotenv==1.2.1
